+++ src/Builder/ShlinkClientBuilder.php
@@ -20,13 +20,13 @@ use Shlinkio\Shlink\SDK\Tags\TagsClientInterface;
 use Shlinkio\Shlink\SDK\Visits\VisitsClient;
 use Shlinkio\Shlink\SDK\Visits\VisitsClientInterface;
 
-class ShlinkClientBuilder implements ShlinkClientBuilderInterface
+readonly class ShlinkClientBuilder implements ShlinkClientBuilderInterface
 {
     public function __construct(
-        private readonly ClientInterface $client,
-        private readonly RequestFactoryInterface $requestFactory,
-        private readonly StreamFactoryInterface $streamFactory,
-        private readonly ?HttpDebuggerInterface $debugger = null,
+        private ClientInterface $client,
+        private RequestFactoryInterface $requestFactory,
+        private StreamFactoryInterface $streamFactory,
+        private ?HttpDebuggerInterface $debugger = null,
     ) {
     }