+++ src/Config/ShlinkConfig.php
@@ -7,29 +7,13 @@ namespace Shlinkio\Shlink\SDK\Config;
 use Shlinkio\Shlink\SDK\Config\Exception\InvalidConfigException;
 use Shlinkio\Shlink\SDK\Http\ApiVersion;
 
-final class ShlinkConfig implements ShlinkConfigInterface
+final readonly class ShlinkConfig implements ShlinkConfigInterface
 {
-    private function __construct(
-        private readonly string $baseUrl,
-        private readonly string $apiKey,
-        private readonly ApiVersion $version,
-    ) {
+    private function __construct(private string $baseUrl, private string $apiKey, private ApiVersion $version)
+    {
     }
 
-    /**
-     * @deprecated Use fromV2BaseUrlAndApiKey instead
-     */
     public static function fromBaseUrlAndApiKey(string $baseUrl, string $apiKey): ShlinkConfigInterface
-    {
-        return self::fromV2BaseUrlAndApiKey($baseUrl, $apiKey);
-    }
-
-    public static function fromV2BaseUrlAndApiKey(string $baseUrl, string $apiKey): ShlinkConfigInterface
-    {
-        return new self($baseUrl, $apiKey, ApiVersion::V2);
-    }
-
-    public static function fromV3BaseUrlAndApiKey(string $baseUrl, string $apiKey): ShlinkConfigInterface
     {
         return new self($baseUrl, $apiKey, ApiVersion::V3);
     }
@@ -66,11 +50,7 @@ final class ShlinkConfig implements ShlinkConfigInterface
             throw InvalidConfigException::forInvalidVersion($version);
         }
 
-        if ($version === '3') {
-            return self::fromV3BaseUrlAndApiKey($baseUrl, $apiKey);
-        }
-
-        return self::fromV2BaseUrlAndApiKey($baseUrl, $apiKey);
+        return self::fromBaseUrlAndApiKey($baseUrl, $apiKey);
     }
 
     public function baseUrl(): string