+++ src/ShortUrls/Model/ShortUrl.php
@@ -6,29 +6,32 @@ namespace Shlinkio\Shlink\SDK\ShortUrls\Model;
 
 use DateTimeImmutable;
 use DateTimeInterface;
-use Shlinkio\Shlink\SDK\Visits\Model\VisitsCount;
+use Shlinkio\Shlink\SDK\Visits\Model\VisitsSummary;
 
-final class ShortUrl
+final readonly class ShortUrl
 {
     /** @deprecated Use $visitsSummary->total instead */
-    public readonly int $visitsCount;
+    public int $visitsCount;
+    /** @deprecated Not returned by Shlink 4.0.0 */
+    public ?DeviceLongUrls $deviceLongUrls;
 
     private function __construct(
-        public readonly string $shortCode,
-        public readonly string $shortUrl,
-        public readonly string $longUrl,
-        public readonly DateTimeInterface $dateCreated,
-        int $visitsCount,
-        public readonly ?string $domain,
-        public readonly ?string $title,
-        public readonly bool $crawlable,
-        public readonly bool $forwardQuery,
-        public readonly array $tags,
-        public readonly ShortUrlMeta $meta,
-        public readonly VisitsCount $visitsSummary,
-        public readonly DeviceLongUrls $deviceLongUrls,
+        public string $shortCode,
+        public string $shortUrl,
+        public string $longUrl,
+        public DateTimeInterface $dateCreated,
+        public ?string $domain,
+        public ?string $title,
+        public bool $crawlable,
+        public bool $forwardQuery,
+        public array $tags,
+        public ShortUrlMeta $meta,
+        public VisitsSummary $visitsSummary,
+        ?DeviceLongUrls $deviceLongUrls,
     ) {
-        $this->visitsCount = $visitsCount;
+        // Not using constructor property promotion here so that we can mark these props as deprecated
+        $this->visitsCount = $visitsSummary->total;
+        $this->deviceLongUrls = $deviceLongUrls;
     }
 
     public static function fromArray(array $payload): self
@@ -41,15 +44,16 @@ final class ShortUrl
             longUrl: $payload['longUrl'] ?? '',
             // @phpstan-ignore-next-line
             dateCreated: DateTimeImmutable::createFromFormat(DateTimeInterface::ATOM, $payload['dateCreated']),
-            visitsCount: $visitsCount,
             domain: $payload['domain'] ?? null,
             title: $payload['title'] ?? null,
             crawlable: $payload['crawlable'] ?? false,
             forwardQuery: $payload['forwardQuery'] ?? false,
             tags: $payload['tags'] ?? [],
             meta: ShortUrlMeta::fromArray($payload['meta'] ?? []),
-            visitsSummary: VisitsCount::fromArrayWithFallback($payload['visitsSummary'] ?? [], $visitsCount),
-            deviceLongUrls: DeviceLongUrls::fromArray($payload['deviceLongUrls'] ?? []),
+            visitsSummary: VisitsSummary::fromArrayWithFallback($payload['visitsSummary'] ?? [], $visitsCount),
+            deviceLongUrls: isset($payload['deviceLongUrls'])
+                ? DeviceLongUrls::fromArray($payload['deviceLongUrls'])
+                : null,
         );
     }
 }