+++ test/Http/HttpClientTest.php
@@ -70,8 +70,8 @@ class HttpClientTest extends TestCase
 
     public static function provideGetRequests(): iterable
     {
-        yield 'no query' => ['/foo/bar', [], 'https://s.test/rest/v2/foo/bar'];
-        yield 'array query' => ['/foo/bar', ['some' => 'thing'], 'https://s.test/rest/v2/foo/bar?some=thing'];
+        yield 'no query' => ['/foo/bar', [], 'https://s.test/rest/v3/foo/bar'];
+        yield 'array query' => ['/foo/bar', ['some' => 'thing'], 'https://s.test/rest/v3/foo/bar?some=thing'];
         yield 'serializable query' => [
             '/foo/bar',
             new class implements ArraySerializable {
@@ -80,7 +80,7 @@ class HttpClientTest extends TestCase
                     return ['foo' => 'bar', 'tags' => ['one', 'two']];
                 }
             },
-            'https://s.test/rest/v2/foo/bar?foo=bar&tags%5B0%5D=one&tags%5B1%5D=two',
+            'https://s.test/rest/v3/foo/bar?foo=bar&tags%5B0%5D=one&tags%5B1%5D=two',
         ];
     }